- Improve significantly the RTMFP introduction service to support server position inside the LAN (with exterior peer) - Fix a metadata bug when publication was running on subscription - Fix a sync issue on initialization of subscriber parameters (in Listener) - Fix a initialization issue with audio AAC stream (without Video) - Add a MediaContainer custom data/metadata pattern - Add FLV custom data and metadata feature (for HTTP direct streaming) - Test and fix onVideo/onAudio feature in streaming - Change dump feature to be filtrable (by protocol for example) - Add thread-safety to log system - Add an Applicaton::path() to simplify access for "folder/path/name" information about the current application - Add a Mona::Server::start signature with explicit Parameters, and move the parameters copies to Invoker in Mona::Server class - Fix a WebSocket.timeout configuration issue - Make Invoker::rootPath virtual pure to force its overload (otherwise dangerous protocol access in executable folder) - Add a "wwwPath" constructor parameter to KulaByteServer class - Remove the two WWWPath and DataPath static variable to be instance variables (more safe) - Fix set RTMP timeout default value like 2x its ping (like for WebSocket) - Add a overloadable onChange method in Mona::Parameters class - Add a String::ToBoolean method to convert String in a boolean type - Parameters::getBoolean use String::ToBoolean now - Fix XMLParser::parseXMLName value returned which was sometimes boolean instead of NULL - Fix peer.protocol name for RTMPE - Add IPAddress::prefixLength() function - Dissociate NETIP/NETPORT exception to know the address part which is invalid - Change client.serverAddress type from string to SocketAddress to be used without error possible in RTMFP introduction/redirection - Code to try to assign automatically client.serverAddress for every protocol - Add a protocol.publicAddress property to resolve in advance the public address of every protocol server - Improve multi-server address exchange/redirection in using protocol.publicAddress - Fix duplicated protocol name issue in protocols - Add static Binary method to convert endian order - Fix minor logs level - Some fixes for Mac - Fix warnings of linux compilation - Few logs and comment addition

+++ MonaBase/include/Mona/Date.h
@@ -42,15 +42,15 @@ public:
 
 
 	enum Type {
-		GMT = 0x7FFFFFFF, /// Special value for offset (Int32 minimum)
-		LOCAL = 0x80000000 /// Special value for offset(Int32 maximum)
+		GMT = (Int32)0x7FFFFFFF, /// Special value for offset (Int32 minimum)
+		LOCAL = (Int32)0x80000000 /// Special value for offset(Int32 maximum)
 	};
 
 	static bool  IsLeapYear(Int32 year) { return (year % 400 == 0) || (!(year & 3) && year % 100); }
 
 	// build a NOW date, not initialized (is null)
 	// /!\ Keep 'Type' to avoid confusion with "build from time" constructor, if a explicit Int32 offset is to set, use Date::setOffset or "build from time" contructor
-	explicit Date(Type offset=LOCAL) : _isDST(false),_year(0), _month(0), _day(0), _weekDay(7),_hour(0), _minute(0), _second(0), _millisecond(0), _changed(false), _offset(offset),_isLocal(true) {}
+	explicit Date(Type offset=LOCAL) : _isDST(false),_year(0), _month(0), _day(0), _weekDay(7),_hour(0), _minute(0), _second(0), _millisecond(0), _changed(false), _offset((Int32)offset),_isLocal(true) {}
 	
 	// build from time
 	explicit Date(Int64 time,Int32 offset=LOCAL) : _isDST(false),_year(0), _month(0), _day(0),  _weekDay(7),_hour(0), _minute(0), _second(0), _millisecond(0), _changed(false), _offset(offset),_isLocal(true), Time(time) {}
